Create a function that simulates a round of Rock Paper Scissors
Created a function that takes a player's choice and the output of the computerPlay function as parameters. This function will determine who wins, if there was a tie, or if an unexpected value is received. The player's choice is currently pre-set and there is not yet a way to receive input.