Complete the rock, paper, scissors committed
Created input prompt for player to make a choice. Play game function has been fixed. console.log have been added to show the player's and the computer's choices, the result of each round, and the result of the whole game. Code has been added to deal with unexpected values as well. This should be everything needed to play rock-paper-scissors in the html window console.